requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.42.1
dj-stripe==2.9.0
Django==5.1.5
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
django-js-asset==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
